requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Django==1.11.4
djangorestframework==3.6.4
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pyee==3.0.3
pytz==2017.2
requests==2.18.4
six==1.10.0
slackclient==1.0.9
slackeventsapi==1.0.0
urllib3==1.22
websocket-client==0.44.0
Werkzeug==0.12.2
